The clusterViewer role is enough to execute get-credentials
GKE provides a predefined IAM role (roles/container.clusterViewer)
that is sufficient to execute `gcloud container clusters
get-credentials` against a cluster, assuming that the user is using
standard Google authentication against their cluster.

In particular, the `getCredentials` permission is *not* needed unless
you are trying to use basic auth or client cert authentication to the
cluster, which no one should be doing.  (Client cert, despite its
appealing name, refers to a *single*, predefined client cert that was
originally used for bootstrapping authentication to a cluster).

For most users, all `gcloud container clusters get-credentials` does
is:
  * Fetch the cluster object from the GKE API, and take note of the
  apiserver's certificate.
  * Write the certificate and the gcloud auth helper into a new
  kubeconfig entry.

When actually authenticating to the cluster, gcloud will use the
configured credential helper to get a Google OAuth access token and
send it to the cluster, where the IAM authorizer will verify it.